+++ src/ast/types.ts
@@ -205,7 +205,7 @@ export interface ClassMethod {
 
 export interface ClassField {
   name: string;
-  fieldType: 'i32' | 'string' | 'string[]' | 'number[]' | 'boolean[]';  // Primitive types and arrays
+  fieldType: 'double' | 'string' | 'string[]' | 'number[]' | 'boolean[]';  // Primitive types and arrays
 }
 
 export interface ClassNode {
